changed entry data model
I don't need to be storing any data of individual entries as Int64, so I reduced some fields to Int32 and Int16

+++ Vridian/Models/LogActivity.swift
@@ -39,15 +39,15 @@ extension Activity {
             for _ in Range(1...amount) {
                 let entry = NSEntityDescription.insertNewObject(forEntityName: "Diary", into: moc) as! Diary
                 
-                entry.actID = Int64(activity.id)
-                entry.catID = Int64(activity.catID)
+                entry.actID = Int32(activity.id)
+                entry.catID = Int16(activity.catID)
                 entry.name = activity.name
                 entry.logged = NSDate() as Date
-                entry.pts = Int64(activity.pts)
+                entry.pts = Int32(activity.pts)
                 entry.kWh = activity.kWh ?? 0
                 entry.mi = activity.mi ?? 0
-                entry.paper = Int64(activity.paper ?? 0)
-                entry.plastic = Int64(activity.plastic ?? 0)
+                entry.paper = Int32(activity.paper ?? 0)
+                entry.plastic = Int32(activity.plastic ?? 0)
                 entry.water = activity.water ?? 0
                 
                 appDelegate?.saveContext()
@@ -57,15 +57,15 @@ extension Activity {
             print("Massive add")
             let entry = Diary(context: moc)
 
-            entry.actID = Int64(activity.id)
-            entry.catID = Int64(activity.catID)
+            entry.actID = Int32(activity.id)
+            entry.catID = Int16(activity.catID)
             entry.name = activity.name
             entry.logged = NSDate() as Date
-            entry.pts = Int64(activity.pts * amount)
+            entry.pts = Int32(activity.pts * amount)
             entry.kWh = Float(activity.kWh ?? 0 * Float(amount))
             entry.mi = Float(activity.mi ?? 0 * Float(amount))
-            entry.paper = Int64(activity.paper ?? 0 * amount)
-            entry.plastic = Int64(activity.plastic ?? 0 * amount)
+            entry.paper = Int32(activity.paper ?? 0 * amount)
+            entry.plastic = Int32(activity.plastic ?? 0 * amount)
             entry.water = Float(activity.water ?? 0 * Float(amount))
 
             appDelegate?.saveContext()
